+++ src/transformers/transformers.ts
@@ -1,14 +1,14 @@
 import { loader } from 'webpack';
 import {
   TransformationDescriptor,
-  TransformationSource
+  TransformationSource,
 } from '../transformation';
 
 // Do not use lambda functions, they won't retain `this` context
 export type TransformationAdapter = (
   this: loader.LoaderContext,
   imagePath: string,
-  transformations: TransformationDescriptor[]
+  transformations: TransformationDescriptor[],
 ) => Promise<TransformationSource[]>;
 
-export type TransformationAdapterPresets = 'thumbor';
+export type TransformationAdapterPresets = 'thumbor' | 'thumbor-docker';